+++ rs/pocket_ic_server/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Subnets with mainnet canister ranges are created with mainnet subnet IDs.
+
 
 
 ## 12.0.0 - 2026-01-23